requirements.txt
aiogram
redis
requests
python-dotenv
